requirements.txt
opencv-python 
pupil-apriltags
pyserial
paho-mqtt